issue #3420 - initial pass at adding azurite env to ci (#3829)
* initial pass at adding azurite env to ci

1. Refactored pre and post docker integration test scripts to take an argument
that identifies which environment to setup/teardown. Almost everything is shared
but each environment can have a single "configure.sh" script to do environment-specific config.
2. Renamed the current "docker" directory to "minio"
3. Introduced a new environment flavor for "azurite"
4. Updated the integration github action workflow to invoke these scripts with an environment argument
from the "environment" test matrix (currently "minio" and "azurite")


* tie the 'set -x' output to the GitHub ACTIONS_STEP_DEBUG


* remove nats.io from azurite test env


* add step to create the azure blob storage container


* move update-server-registry-resource.sh to common


* use 'default' for the azure storageProvider to match the existing tests


* leave the azure import tests disabled for now


* one more fix

Locally docker compose uses [dirName]-[serviceName]-[number]
But in the CI its using [dirName]_[serviceName]_[number] (underscore instead of dash).
I updated the scripts to match either.


* more cleanup

remove the no-longer-used test-import files
and minor tweaks in post-integration-test-docker.sh


* issue #3304 - favor Exceptions over Errors in ExampleRequestProcessor

The method signature indicates that Exception is thrown, but the
implementation was throwing AssertionErrors instead.

I made some other minor changes (mostly style preference stuff) along
the way and included those in the commit.


* Set the fhirVersion from ExampleRequestProcessor

If the request uses "application/fhir+json", the default server config
will treat that like version 4.0.
That works for most resource types, but fails for both the resource
types introduced in FHIR R4B as well as for Evidence and
EvidenceVariable which have backwards-incompatible changes in R4B.


* address review feedback

1. added test to restrict the input argument for pre and post docker test scripts
   and store the validated string into a new var named 'subdir'
2. moved pre-test log gathering into new script common/gather-pretest-logs.sh
3. changed existing gather-logs.sh file name to gather-posttest-logs.sh
4. changed '_' to '-' in a couple other common scripts for consistency


* rename docker folder and images

"ibmcom/ibm-fhir-*" -> "linuxforhealth/fhir-*"

I kept the ibm-fhir-server name in two places:
1. the migration tests (so that we can continue using a mix of old and new); and
2. the security workflow (because I havn't touched this one much and it probably needs some dedicated attention)

We've decided to use GitHub Packages (ghcr) for our docker images moving forward,
so we'll have more release automation work to make that happen.

+++ build/post-integration-test-docker.sh
@@ -4,13 +4,28 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 ###############################################################################
-set -x
+
+# Exit the script if any commands fail
+set -e
+# Print each command before executing it
+#set -x
+# This allows subshells to inheret the options above
+export SHELLOPTS
 
 # The full path to the directory of this script, no matter where its called from
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
-WORKSPACE="$( dirname "${DIR}" )"
+export WORKSPACE="$( dirname "${DIR}" )"
 
-echo "Performing integration test post-processing..."
+if [ -z "${1}" ]; then
+  echo "This script requires an argument that points to one of the integration environment folders."
+  exit 1
+elif [ -z "$(find ${DIR} -maxdepth 1 -name ${1} -type d)" ]; then
+  echo "'${1}' is not a valid argument. This script requires an argument that points to one of the integration environment folders."
+  exit 1
+fi
+
+subdir="${1}"
+echo "Bringing down environment for the ${subdir} fhir-server integration tests..."
 
 # Log output locations
 it_results=${WORKSPACE}/integration-test-results
@@ -21,9 +36,9 @@ rm -rf ${it_results} 2>/dev/null
 mkdir -p ${it_results}/server-logs
 mkdir -p ${it_results}/fhir-server-test
 
-containerId=$(docker ps -a | grep fhir | cut -d ' ' -f 1)
-if [[ -z "${containerId}" ]]; then
-    echo "Warning: Could not find fhir container!!!"
+containerId=$(docker ps -a | grep ${subdir}[-_]fhir-server | cut -d ' ' -f 1)
+if [ -z "${containerId}" ]; then
+    echo "Warning: Could not find fhir-server container!!!"
 else
     echo "fhir container id: $containerId"
 
@@ -35,12 +50,16 @@ else
 fi
 
 echo "Gathering integration test output"
-cp -r ${WORKSPACE}/fhir-server-test/target/surefire-reports/* ${it_results}/fhir-server-test
+if [ -e ${WORKSPACE}/fhir-server-test/target/surefire-reports ]; then
+    cp -r ${WORKSPACE}/fhir-server-test/target/surefire-reports/* ${it_results}/fhir-server-test
+fi
+
+# Source the tenant1 datastore variables; not strictly needed here
+# but avoids "variable is not set" docker compose warnings
+. ${WORKSPACE}/build/common/set-tenant1-datastore-vars.sh
 
 echo "Bringing down the fhir server docker container(s)..."
-cd ${DIR}/docker
-docker-compose down
+cd ${DIR}/${subdir}
+docker compose down
 
 echo "Integration test post-processing completed!"
-
-exit 0
